requirement.txt
fastmcp 
hyperliquid-python-sdk 
python-dotenv 
eth-account